reduce DTO to double (was long double)

+++ debug/anglib.hpp
@@ -14,7 +14,7 @@ namespace anglib
 
 	struct proxy_imag
 	{
-		long double sec;
+		double sec;
 		long dh;
 		short min;
 	};
